requirements.txt
# Core dependencies — pinned for reproducible builds
# Dev: pytest==8.3.5
google-auth==2.48.0
google-api-python-client==2.190.0
google-auth-httplib2==0.3.0
sshtunnel==0.4.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
requests==2.32.5
flask==3.1.2
paramiko==3.5.1
fastapi==0.129.0
uvicorn==0.41.0
pydantic==2.12.5
